+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased Changes]
 
 - add .lazy_attr_accessor
+- .lazy_attr_reader and .lazy_attr_accessor return the same value as their normal counterparts
 
 ## 0.1.0